requirements.txt
torch 
transformers 
langchain 
sentence-transformers 
faiss-cpu 
chromadb 
streamlit